requirements.txt
asgiref==3.8.1
Django==5.1.4
django-json==0.27
django-rest-framework==0.1.0
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
sqlparse==0.5.3
